+++ linuxdoom-1.10/i_net.c
@@ -53,10 +53,10 @@ rcsid[] = "$Id: m_bbox.c,v 1.1 1997/02/03 22:45:10 b1 Exp $";
 
 // For some odd reason...
 #define ntohl(x) \
-        ((unsigned long int)((((unsigned long int)(x) & 0x000000ffU) << 24) | \
-                             (((unsigned long int)(x) & 0x0000ff00U) <<  8) | \
-                             (((unsigned long int)(x) & 0x00ff0000U) >>  8) | \
-                             (((unsigned long int)(x) & 0xff000000U) >> 24)))
+        ((uint32_t)((((uint32_t)(x) & 0x000000ffU) << 24) | \
+                    (((uint32_t)(x) & 0x0000ff00U) <<  8) | \
+                    (((uint32_t)(x) & 0x00ff0000U) >>  8) | \
+                    (((uint32_t)(x) & 0xff000000U) >> 24)))
 
 #define ntohs(x) \
         ((unsigned short int)((((unsigned short int)(x) & 0x00ff) << 8) | \
